+++ packages/components-vue/src/examples/index.ts
@@ -1,7 +1,11 @@
 export { default as AccordionExample } from './AccordionExample.vue';
+export { default as SheetExample } from './SheetExample.vue';
 export { default as AGGridExample } from './AGGridExample.vue';
 export { default as AGGridExampleStorefront } from './AGGridExampleStorefront.vue';
 export { default as ButtonExampleForm } from './ButtonExampleForm.vue';
+export { default as ButtonExampleFormAttribute } from './ButtonExampleFormAttribute.vue';
+export { default as ButtonPureExampleForm } from './ButtonPureExampleForm.vue';
+export { default as ButtonPureExampleFormAttribute } from './ButtonPureExampleFormAttribute.vue';
 export { default as CanvasExample } from './CanvasExample.vue';
 export { default as CarouselExampleDynamicSlides } from './CarouselExampleDynamicSlides.vue';
 export { default as CarouselExampleFocusOnCenterSlide } from './CarouselExampleFocusOnCenterSlide.vue';
@@ -10,6 +14,7 @@ export { default as CarouselExampleJumpToSlide } from './CarouselExampleJumpToSl
 export { default as CheckboxExample } from './CheckboxExample.vue';
 export { default as CheckboxExampleControlled } from './CheckboxExampleControlled.vue';
 export { default as FlyoutExample } from './FlyoutExample.vue';
+export { default as FlyoutExampleForm } from './FlyoutExampleForm.vue';
 export { default as FlyoutMultilevelExample } from './FlyoutMultilevelExample.vue';
 export { default as FlyoutMultilevelExampleActiveIdentifier } from './FlyoutMultilevelExampleActiveIdentifier.vue';
 export { default as FlyoutMultilevelExampleCustomContent } from './FlyoutMultilevelExampleCustomContent.vue';