+++ packages/components/src/components/button-pure/button-pure.tsx
@@ -1,23 +1,25 @@
+import { AttachInternals, Component, Element, Host, type JSX, Listen, Prop, Watch, h } from '@stencil/core';
+import type { BreakpointCustomizable, PropTypes, SelectedAriaAttributes, Theme } from '../../types';
 import {
   ALIGN_LABELS,
   AllowedTypes,
-  attachComponentCss,
   BUTTON_ARIA_ATTRIBUTES,
   BUTTON_TYPES,
+  TEXT_SIZES,
+  THEMES,
+  TYPOGRAPHY_TEXT_WEIGHTS,
+  attachComponentCss,
   getPrefixedTagNames,
   hasPropValueChanged,
   hasVisibleIcon,
   improveButtonHandlingForCustomElement,
   isDisabledOrLoading,
-  TEXT_SIZES,
-  TYPOGRAPHY_TEXT_WEIGHTS,
-  THEMES,
   validateProps,
   warnIfDeprecatedPropValueIsUsed,
   warnIfParentIsPTextAndIconIsNone,
 } from '../../utils';
-import type { BreakpointCustomizable, PropTypes, SelectedAriaAttributes, Theme } from '../../types';
-import { Component, Element, h, Host, type JSX, Listen, Prop } from '@stencil/core';
+import { LoadingMessage, loadingId } from '../common/loading-message/loading-message';
+import { getComponentCss } from './button-pure-styles';
 import {
   type ButtonPureAlignLabel,
   type ButtonPureAlignLabelDeprecated,
@@ -29,8 +31,6 @@ import {
   getButtonPureAriaAttributes,
   warnIfIsLoadingAndIconIsNone,
 } from './button-pure-utils';
-import { getComponentCss } from './button-pure-styles';
-import { LoadingMessage, loadingId } from '../common/loading-message/loading-message';
 
 const propTypes: PropTypes<typeof ButtonPure> = {
   type: AllowedTypes.oneOf<ButtonPureType>(BUTTON_TYPES),
@@ -49,6 +49,7 @@ const propTypes: PropTypes<typeof ButtonPure> = {
   stretch: AllowedTypes.breakpoint('boolean'),
   theme: AllowedTypes.oneOf<Theme>(THEMES),
   aria: AllowedTypes.aria<ButtonPureAriaAttribute>(BUTTON_ARIA_ATTRIBUTES),
+  form: AllowedTypes.string,
 };
 
 /**
@@ -57,6 +58,7 @@ const propTypes: PropTypes<typeof ButtonPure> = {
 @Component({
   tag: 'p-button-pure',
   shadow: { delegatesFocus: true },
+  formAssociated: true,
 })
 export class ButtonPure {
   @Element() public host!: HTMLElement;
@@ -65,7 +67,7 @@ export class ButtonPure {
   @Prop() public type?: ButtonPureType = 'submit';
 
   /** The name of the button, submitted as a pair with the button's value as part of the form data, when that button is used to submit the form. */
-  @Prop() public name?: string;
+  @Prop({ reflect: true }) public name?: string;
 
   /** Defines the value associated with the button's name when it's submitted with the form data. This value is passed to the server in params when the form is submitted using this button. */
   @Prop() public value?: string;
@@ -112,6 +114,12 @@ export class ButtonPure {
   /** Add ARIA attributes. */
   @Prop() public aria?: SelectedAriaAttributes<ButtonPureAriaAttribute>;
 
+  /** The id of a form element the button should be associated with. */
+  @Prop({ reflect: true }) public form?: string;
+  // In the React wrapper, all props are synced as properties on the element ref, so reflecting "form" as an attribute ensures it is properly handled in the form submission process.
+
+  @AttachInternals() private internals: ElementInternals;
+
   private initialLoading: boolean = false;
 
   private get isDisabledOrLoading(): boolean {
@@ -123,6 +131,24 @@ export class ButtonPure {
   public onClick(e: MouseEvent): void {
     if (this.isDisabledOrLoading) {
       e.stopPropagation();
+      return;
+    }
+
+    if (this.form && this.internals?.form) {
+      e.preventDefault();
+      if (this.type === 'submit') {
+        // Submitter is null because the button can't be passed from the shadow DOM https://github.com/WICG/webcomponents/issues/814
+        this.internals?.form.requestSubmit();
+      } else if (this.type === 'reset') {
+        this.internals?.form.reset();
+      }
+    }
+  }
+
+  @Watch('value')
+  public onValueChange(newValue: string): void {
+    if (this.form) {
+      this.internals?.setFormValue(newValue);
     }
   }
 
@@ -131,6 +157,9 @@ export class ButtonPure {
   }
 
   public componentWillLoad(): void {
+    if (this.form) {
+      this.internals?.setFormValue(this.value);
+    }
     this.initialLoading = this.loading;
   }
 
@@ -145,13 +174,15 @@ export class ButtonPure {
   }
 
   public componentDidLoad(): void {
-    improveButtonHandlingForCustomElement(
-      this.host,
-      () => this.type,
-      () => this.isDisabledOrLoading,
-      () => this.name,
-      () => this.value
-    );
+    if (!this.form) {
+      improveButtonHandlingForCustomElement(
+        this.host,
+        () => this.type,
+        () => this.isDisabledOrLoading,
+        () => this.name,
+        () => this.value
+      );
+    }
   }
 
   public render(): JSX.Element {