+++ packages/components/src/components/button/button.spec.ts
@@ -4,6 +4,11 @@ const initComponent = (): Button => {
   const component = new Button();
   component.host = document.createElement('p-button');
   component.host.attachShadow({ mode: 'open' });
+  component['internals'] = {
+    setFormValue: jest.fn(),
+    form: { requestSubmit: jest.fn(), reset: jest.fn() } as unknown as HTMLFormElement,
+  } as unknown as ElementInternals;
+  component.form = 'some-form';
   return component;
 };
 
@@ -45,5 +50,109 @@ describe('componentWillLoad', () => {
     component.loading = false;
     component.componentWillLoad();
     expect(component['initialLoading']).toBe(false);
+  });
+
+  it('should call setFormValue() if form prop is defined', () => {
+    const component = initComponent();
+    const value = 'some-value';
+    component.form = 'some-form';
+    component.value = value;
+    component.componentWillLoad();
+    expect(component['internals'].setFormValue).toHaveBeenCalledWith(value);
+  });
+
+  it('should not call setFormValue() if form prop is undefined', () => {
+    const component = initComponent();
+    component.form = undefined;
+    component.componentWillLoad();
+    expect(component['internals'].setFormValue).not.toHaveBeenCalled();
+  });
+});
+
+describe('onValueChange()', () => {
+  it('should call setFormValue() if form prop is defined', () => {
+    const component = initComponent();
+    const value = 'some-value';
+    component.form = 'some-form';
+    component.onValueChange(value);
+    expect(component['internals'].setFormValue).toHaveBeenCalledWith(value);
+  });
+
+  it('should not call setFormValue() if form prop is undefined', () => {
+    const component = initComponent();
+    component.form = undefined;
+    component.onValueChange('some-value');
+    expect(component['internals'].setFormValue).not.toHaveBeenCalled();
+  });
+});
+
+describe('onClick()', () => {
+  let mockEvent: MouseEvent;
+  let component: Button;
+
+  beforeEach(() => {
+    mockEvent = new MouseEvent('click');
+    jest.spyOn(mockEvent, 'stopPropagation');
+    jest.spyOn(mockEvent, 'preventDefault');
+
+    component = initComponent();
+  });
+
+  afterEach(() => {
+    jest.clearAllMocks();
+  });
+
+  it('should stop propagation if the component is disabled or loading', () => {
+    component.loading = true;
+
+    component.onClick(mockEvent);
+
+    expect(mockEvent.stopPropagation).toHaveBeenCalled();
+    expect(mockEvent.preventDefault).not.toHaveBeenCalled();
+    expect(component['internals'].form.requestSubmit).not.toHaveBeenCalled();
+    expect(component['internals'].form.reset).not.toHaveBeenCalled();
+  });
+
+  it('should call requestSubmit for type "submit"', () => {
+    component.loading = false;
+    component.type = 'submit';
+
+    component.onClick(mockEvent);
+
+    expect(mockEvent.preventDefault).toHaveBeenCalled();
+    expect(component['internals'].form.requestSubmit).toHaveBeenCalled();
+    expect(component['internals'].form.reset).not.toHaveBeenCalled();
+  });
+
+  it('should call reset for type "reset"', () => {
+    component.loading = false;
+    component.type = 'reset';
+
+    component.onClick(mockEvent);
+
+    expect(mockEvent.preventDefault).toHaveBeenCalled();
+    expect(component['internals'].form.requestSubmit).not.toHaveBeenCalled();
+    expect(component['internals'].form.reset).toHaveBeenCalled();
+  });
+
+  it('should not call any form actions for unhandled types', () => {
+    component.loading = false;
+    component.type = 'button';
+
+    component.onClick(mockEvent);
+
+    expect(component['internals'].form.requestSubmit).not.toHaveBeenCalled();
+    expect(component['internals'].form.reset).not.toHaveBeenCalled();
+  });
+
+  it('should not call any form actions for missing form id', () => {
+    component.form = null;
+    component.loading = false;
+    component.type = 'submit';
+
+    component.onClick(mockEvent);
+
+    expect(component['internals'].form.requestSubmit).not.toHaveBeenCalled();
+    expect(component['internals'].form.reset).not.toHaveBeenCalled();
   });
 });