+++ packages/components/src/components/carousel/carousel.spec.ts
@@ -94,8 +94,8 @@ describe('connectedCallback', () => {
 });
 
 describe('componentWillLoad', () => {
-  it('should call parseJSON() with correct parameters and set slidesPerPage', () => {
-    const spy = jest.spyOn(breakpointCustomizableUtils, 'parseJSON').mockReturnValueOnce(5).mockReturnValueOnce(10);
+  it('should call parseJSON() in private "parsedSlidesPerPage()" with correct parameters', () => {
+    const spy = jest.spyOn(breakpointCustomizableUtils, 'parseJSON').mockReturnValue(5);
     const component = new Carousel();
     component.host = document.createElement('p-carousel');
     component.slidesPerPage = 2;
@@ -103,7 +103,8 @@ describe('componentWillLoad', () => {
     component.componentWillLoad();
     expect(spy).toHaveBeenCalledWith(2);
 
-    expect(component.slidesPerPage).toBe(5);
+    expect(component.slidesPerPage).toBe(2);
+    expect((component as any).parsedSlidesPerPage).toBe(5);
   });
 
   it('should call this.updateSlidesAndPagination()', () => {