+++ packages/components/src/components/table/table-row/table-row-styles.ts
@@ -1,6 +1,10 @@
-import { getCss } from '../../../utils';
 import { addImportantToEachRule, getTransition, hostHiddenStyles, hoverMediaQuery } from '../../../styles';
-import { cssVariableTableBorderColor, cssVariableTableHoverColor } from '../table/table-styles';
+import { getCss } from '../../../utils';
+import {
+  cssVariableTableBorderColor,
+  cssVariableTableBorderWidth,
+  cssVariableTableHoverColor,
+} from '../table/table-styles';
 
 export const getComponentCss = (): string => {
   return getCss({
@@ -8,8 +12,8 @@ export const getComponentCss = (): string => {
       ':host': {
         display: 'table-row',
         ...addImportantToEachRule({
-          borderTop: `1px solid var(${cssVariableTableBorderColor})`,
-          borderBottom: `1px solid var(${cssVariableTableBorderColor})`,
+          borderTop: `var(${cssVariableTableBorderWidth},1px) solid var(${cssVariableTableBorderColor})`,
+          borderBottom: `var(${cssVariableTableBorderWidth},1px) solid var(${cssVariableTableBorderColor})`,
           transition: getTransition('background'),
           ...hostHiddenStyles,
           ...hoverMediaQuery({